refactor(message api): reorganize Anthropic router into layered architecture (#425)
