--- model_gateway/src/routers/anthropic/messages/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod non_streaming;
-pub(crate) mod streaming;
-pub(crate) mod tools;
