+++ model_gateway/src/routers/anthropic/mod.rs
@@ -1,8 +1,11 @@
 pub(crate) mod context;
-pub(crate) mod handler;
-pub mod messages;
+pub(crate) mod mcp;
 pub mod models;
+pub(crate) mod non_streaming;
 mod router;
+pub(crate) mod sse;
+pub(crate) mod streaming;
 pub(crate) mod utils;
+pub(crate) mod worker;
 
 pub use router::AnthropicRouter;
